requirements.txt
asgiref==3.2.7 
django==3.0.5 
pytz==2020.1 
sqlparse==0.3.1
djangorestframework==3.11.0
